+++ spec/oracle_i386_spec.rb
@@ -14,17 +14,6 @@ describe 'java::oracle_i386' do
     expect(chef_run).to include_recipe('java::notify')
   end
 
-  it 'should notify of jdk-version-change' do
-    pending 'Testing LWRP use is not required at this time, this is tested post-converge.'
-    expect(chef_run.jdk_ark('jdk-alt')).to notify('log[jdk-version-changed]')
-    this_should_not_get_executed
-  end
-
-  it 'should configure a java_oracle_install[jdk] resource' do
-    pending 'Testing LWRP use is not required at this time, this is tested post-converge.'
-    this_should_not_get_executed
-  end
-
   describe 'conditionally includes set attributes' do
     context 'when java_home is set' do
       let(:chef_run) do