+++ spec/spec_helper.rb
@@ -1,7 +1,9 @@
 $LOAD_PATH.unshift(File.join(File.dirname(__FILE__), '..', 'libraries'))
-require 'helpers'
 require 'chefspec'
 require 'chefspec/berkshelf'
+
+require_relative '../libraries/adopt_openjdk_helpers.rb'
+require_relative '../libraries/helpers.rb'
 
 RSpec.configure do |config|
   config.file_cache_path = File.join(Dir.tmpdir, 'chefspec') if config.respond_to?(:file_cache_path)