+++ test/fixtures/cookbooks/test/recipes/adoptopenjdk-10-openj9.rb
@@ -1,6 +1,0 @@
-node.default['java']['install_flavor'] = 'adoptopenjdk'
-node.default['java']['jdk_version'] = '10'
-
-include_recipe 'test::base'
-include_recipe 'java::default'
-include_recipe 'test::java_cert'